pyproject.toml
[project]
name = "258839R-IT5437-CV-Assignment-2"
version = "0.1.0"
description = "IT5437 Assignment 2 - Fitting and Alignment"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "ipympl>=0.10.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "python-docx>=1.2.0",
    "scikit-image>=0.26.0",
    "scipy>=1.17.1",
]
